+++ Examples/Resources/AuditSetting/1-AuditSetting_AuditVolumneNtfsConfig.ps1
@@ -1,4 +1,4 @@
-configuration AuditSystem_AuditVolumneNtfsConfig
+configuration AuditSetting_AuditVolumneNtfsConfig
 {
     Import-DscResource -ModuleName AuditSystemDsc