+++ Examples/Resources/AuditSetting/2-AuditSetting_AuditUsersPasswordNotRequiredConfig.ps1
@@ -1,0 +1,15 @@
+configuration AuditSetting_AuditUsersPasswordNotRequiredConfig
+{
+    Import-DscResource -ModuleName AuditSystemDsc
+
+    node localhost
+    {
+        AuditSetting LocalAccountWithoutPassword
+        {
+            Query = "SELECT * FROM Win32_UserAccount WHERE Disabled = $false"
+            Property = "PasswordRequired"
+            DesiredValue = $true
+            Operator = '-eq'
+        }
+    }
+}